update pbr project

+++ Demos/assets/shaders/irradiance.glsl
@@ -47,7 +47,7 @@ void main()
     up         = normalize(cross(N, right));
        
     float sampleDelta = 0.025;
-    float nrSamples = 0.0;
+    unsigned int nrSamples = 0u;
     for (float phi = 0.0; phi < 2.0 * PI; phi += sampleDelta)
     {
         for (float theta = 0.0; theta < 0.5 * PI; theta += sampleDelta)
